move helper functions and add new loss
